Extract classic player controls specs into separate file
