+++ entry_types/scrolled/package/spec/frontend/PlayerControls-spec.js
@@ -10,7 +10,6 @@ import {renderInEntry} from 'support';
 describe('PlayerControls', () => {
   useFakeTranslations({
     'pageflow_scrolled.public.player_controls.play': 'Play',
-    'pageflow_scrolled.public.player_controls.quality': 'Quality',
     'pageflow_scrolled.public.player_controls.text_tracks': 'Text Tracks'
   });
 
@@ -68,41 +67,6 @@ describe('PlayerControls', () => {
       fireEvent.click(getByText('English'));
 
       expect(listener).toHaveBeenCalledWith('en');
-    });
-  });
-
-  describe('classic player controls variant', () => {
-    it('supports rendering and handling events for quality menu items', () => {
-      const listener = jest.fn();
-      const {getByTitle, getByText} = renderInEntry(
-        <ClassicPlayerControls
-          qualityMenuItems={[
-            {label: '1080p', value: 'fullhd'},
-            {label: '720p', value: 'medium'},
-          ]}
-          onQualityMenuItemClick={listener} />
-      );
-
-      fireEvent.click(getByTitle('Quality'));
-      fireEvent.click(getByText('1080p'));
-
-      expect(listener).toHaveBeenCalledWith('fullhd');
-    });
-
-    it('renders control bar by default', () => {
-      const {queryByLabelText} = renderInEntry(
-        <ClassicPlayerControls />
-      );
-
-      expect(queryByLabelText('Play')).not.toBeNull();
-    });
-
-    it('supports hiding control bar', () => {
-      const {queryByLabelText} = renderInEntry(
-        <ClassicPlayerControls hideControlBar={true} />
-      );
-
-      expect(queryByLabelText('Play')).toBeNull();
     });
   });
 });
